Remove distribute (deprecated)
- use latest version of setuptools, 1.4.2, instead.

Distribute has now been merged into the setuptools codebase.
