+++ MANIFEST.default
@@ -1,4 +1,4 @@
-include distribute_setup.py
+include ez_setup.py
 include python/pyhrf/datafiles/*
 include README.md
 prune script/misc
