+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include distribute_setup.py
+include ez_setup.py
 include python/pyhrf/datafiles/*
